+++ node/src/fs/handle/index.ts
@@ -15,10 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export type * as types from "./types";
-
-export * from "./fs";
-export * from "./client/Client";
-export * from "./errors";
-export * as siyuan from "./utils/siyuan";
-export { default as CONSTANTS } from "./constants";
+export * from "./SiyuanFileSystemDirectoryHandle";
+export * from "./SiyuanFileSystemFileHandle";
+export * from "./SiyuanFileSystemHandle";
+export * from "./SiyuanFileSystemWritableFileStream";