+++ node/src/fs/path-browserify.d.ts
@@ -15,10 +15,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export type * as types from "./types";
-
-export * from "./fs";
-export * from "./client/Client";
-export * from "./errors";
-export * as siyuan from "./utils/siyuan";
-export { default as CONSTANTS } from "./constants";
+declare module "path-browserify" {
+    // export * from "path";
+    import path from "node:path";
+    export default path;
+}